Move visitMarkable and visitMarker up to TreeVisitor
