+++ rewrite-core/src/main/java/org/openrewrite/TreeVisitor.java
@@ -18,8 +18,11 @@ package org.openrewrite;
 import io.micrometer.core.instrument.Metrics;
 import io.micrometer.core.instrument.Timer;
 import org.openrewrite.internal.lang.Nullable;
+import org.openrewrite.marker.Marker;
+import org.openrewrite.marker.Markers;
 
 import java.util.ArrayList;
+import java.util.Collection;
 import java.util.List;
 import java.util.function.BiFunction;
 
@@ -182,5 +185,27 @@ public abstract class TreeVisitor<T extends Tree, P> {
     protected final <T2 extends T> T2 visitAndCast(@Nullable Tree tree, P p) {
         //noinspection unchecked
         return (T2) visit(tree, p);
+    }
+
+    @Incubating(since = "7.2.0")
+    public Markers visitMarkers(Markers markers, P p) {
+        Collection<? extends Marker> originalMarkers = markers.entries();
+        List<Marker> visited = new ArrayList<>();
+        boolean anyChanged = false;
+        for(Marker originalMarker : originalMarkers) {
+            Marker visitedMarker = visitMarker(originalMarker, p);
+            visited.add(visitedMarker);
+            anyChanged = anyChanged || (visitedMarker != originalMarker);
+        }
+        if(anyChanged) {
+            return Markers.build(visited);
+        }
+        return markers;
+    }
+
+    @Incubating(since = "7.2.0")
+    public <M extends Marker> M visitMarker(Marker marker, P p) {
+        //noinspection unchecked
+        return (M) marker;
     }
 }
